requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
pymongo==4.9.0
motor==3.6.0
python-dotenv==1.0.1
pydantic==2.10.3
pydantic-settings==2.6.1
python-multipart==0.0.17
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
boto3==1.35.83
openai>=1.55.3,<2.0.0
numpy==1.26.4
langchain==0.3.13
langchain-openai==0.2.13
langchain-community==0.3.13
langchain-mongodb==0.3.0
langchain-text-splitters==0.3.5
langchain-anthropic==0.2.4
anthropic==0.67.0
PyPDF2==3.0.1
pypdf==5.1.0
sentence-transformers==2.7.0
transformers==4.40.0
torch==2.2.0
python-docx==1.1.2
openpyxl==3.1.5
python-pptx==1.0.2
chardet==5.2.0
beautifulsoup4==4.12.3
requests==2.32.3
langchain-core==0.3.29
unstructured==0.18.14
kafka-python==2.0.2
firecrawl-py==4.3.6
